@@ -2,11 +2,12 @@
   "name": "@mindarchitect-ngx-libraries/ngx-avatar-lib",
   "version": "0.0.4",
   "peerDependencies": {
-      "@angular/common": "^15.0.0",
-      "@angular/core": "^15.0.0",
-      "bootstrap": "^5.2.3"
+    "@angular/common": "^16.0.0",
+    "@angular/core": "^16.0.0",
+    "@angular/material": "^16.0.0",
+    "bootstrap": "^5.2.3"
   },
   "dependencies": {
-      "tslib": "^2.5.0"
+    "tslib": "^2.5.0"
   }
 }
