@@ -2,7 +2,10 @@
  * Public API Surface of ngx-avatar-lib
  */
 
-export * from './lib/ngx-avatar-lib.configuration';
+export * from './lib/configuration/ngx-avatar-lib.configuration';
+export * from './lib/configuration/ngx-avatar-lib-configuration.provider';
+export * from './lib/configuration/default-ngx-avatar-lib-configuration-provider';
+
 export * from './lib/services/ngx-avatar-lib.service';
 export * from './lib/components/ngx-avatar-lib/ngx-avatar-lib.component';
 export * from './lib/ngx-avatar-lib.module';
